backend/requirements.txt
flask
flask-sqlalchemy
flask-cors
flask-jwt-extended
werkzeug
apscheduler
python-dotenv
gunicorn